fhs/usr/share/python/syncwerk/restapi/requirements.txt
aenum==2.2.3
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.15
django-appconf==1.0.2
django-compressor==2.2
django-constance==2.1.0
django-cors-headers==2.1.0
django-formtools==2.1
django-picklefield==1.0.0
# django-post-office==3.0.4
django-ranged-response==0.2.0
django-rest-swagger==2.1.2
django-simple-captcha==0.5.6
django-statici18n==1.7.0
django-webpack-loader==0.6.0
djangorestframework==3.9.1
drf-yasg==1.13.0
enum34==1.1.10
et-xmlfile==1.0.1
flup==1.0.2
futures==3.1.1
gunicorn==19.8.1
idna==2.6
inflection==0.3.1
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
jsonfield==2.0.2
MarkupSafe==1.1.0
mysqlclient==1.3.12
oauthlib==2.1.0
openapi-codec==1.3.2
openpyxl==2.3.0
Pillow==5.1.0
python-dateutil==2.2
python-memcached==1.59
pytz==2018.3
qrcode==5.3
rcssmin==1.0.6
requests==2.18.4
requests-oauthlib==1.0.0
rjsmin==1.0.12
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.88
simplejson==3.16.0
six==1.11.0
git+https://github.com/syncwerk/syncwerk-server-python.git
git+https://github.com/syncwerk/syncwerk-server-python-ccnet.git
git+https://github.com/syncwerk/syncwerk-server-python-django-registration.git
git+https://github.com/syncwerk/syncwerk-server-python-django-shibboleth.git
git+https://github.com/syncwerk/syncwerk-server-python-django-termsandconditions.git
git+https://github.com/syncwerk/syncwerk-server-python-object-storage.git
git+https://github.com/syncwerk/syncwerk-server-python-rpc-bindings.git
git+https://github.com/syncwerk/syncwerk-server-python-rpc-client.git
git+https://github.com/syncwerk/syncwerk-server-python-webdav.git
git+https://github.com/syncwerk/syncwerk-server-python-libevent.git
git+https://github.com/syncwerk/syncwerk-server-python-django-post-office.git
uritemplate==3.0.0
urllib3==1.22
sqlalchemy==1.3.2
sqlparse==0.3.0
elasticsearch==7.10.1
python-magic==0.4.18

